requirements.txt
numpy==1.24.1
pandas==2.0.1
Pillow==9.3.0
spectral==0.23.1
effdet==0.4.1
ensemble-boxes==1.0.6
albumentations==1.3.0
alive-progress==3.1.3
matplotlib==3.7.1
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.9.1
timm==0.9.2
fastcore==1.3.20
pytorch-lightning==2.0.2